Add terraform for aws ec2 + s3
Refactor call_polly to first check the bucket for the mp3, then
call Polly if it is not there. Also added a helper function to
write to a temp file which is the asset sent to Discord. The
file is deleted after sending. This might need to be more robust in the
future.